notification module awesome

+++ lib/tasks/notification.rake
@@ -3,25 +3,29 @@ namespace :notification do
   task send_message: :environment do
     gcm = GCM.new("AIzaSyDi8ducYJNid4NQUKm04HzmDrdScGqO1NE")
     registration_id = Device.all().map(&:token)
-    notification = Notification.today.first
-    options = {
-      "collapse_key" => "updated_state",
-      "delay_while_idle" => true,
-      :data => {
-        "message" => notification.message,
-        "summaryText" => notification.message,
-        "title" => notification.title
+    notification = Notification.today.where(is_published: false).first
+    if (notification)
+      options = {
+        "collapse_key" => "updated_state",
+        "delay_while_idle" => true,
+        :data => {
+          "message" => notification.message,
+          "summaryText" => notification.message,
+          "title" => notification.title
+        }
       }
-    }
 
-    if notification.picture.url != "/pictures/original/missing.png"
-      options[:data][:style] = "picture"
-      options[:data][:picture] = notification.picture.url
+      if notification.picture.url != "/pictures/original/missing.png"
+        options[:data][:style] = "picture"
+        options[:data][:picture] = notification.picture.url
+      end
+      response = gcm.send_notification(registration_id, options)
+      notification.is_published = true;
+      notification.save
+      if (response[:not_registered_ids])
+        Device.where(token: response[:not_registered_ids]).destroy_all
+      end
     end
-    response = gcm.send_notification(registration_id, options)
-    notification.is_published = true;
-    notification.save
-    Device.where(token: response[:not_registered_ids]).destroy_all
   end
 
   desc "TODO"
